+++ tests/spec/defaultResponse/schema.ts
@@ -203,7 +203,11 @@ export class HttpClient<SecurityDataType = unknown> {
 /**
  * @title Swagger Petstore
  * @version 1.0.0
+ * @license Apache 2.0 (https://www.apache.org/licenses/LICENSE-2.0.html)
+ * @termsOfService http://swagger.io/terms/
  * @baseUrl http://petstore.swagger.io/api
+ * @contact Swagger API Team <apiteam@swagger.io> (http://swagger.io)
+ *
  * A sample API that uses a petstore as an example to demonstrate features in the swagger-2.0 specification
  */
 export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDataType> {